🚚 resource: relocate UI related code to Restaurants Component
